# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moling API",
    "description": "墨灵 — AI 小说创作平台后端服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/import": {
      "post": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Submit Import",
        "description": "提交导入任务。\n\n- 粘贴文本：传入 text 参数\n- 上传文件：传入 file 参数（multipart/form-data）\n- 对应 Phase 0 的采集与分章",
        "operationId": "submit_import_api_v1_projects__project_id__import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "粘贴文本导入",
              "title": "Text"
            },
            "description": "粘贴文本导入"
          },
          {
            "name": "split_strategies",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "拆分策略，逗号分隔，如 'chapter_regex,paragraph'",
              "title": "Split Strategies"
            },
            "description": "拆分策略，逗号分隔，如 'chapter_regex,paragraph'"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_import_api_v1_projects__project_id__import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "List Import Jobs",
        "description": "获取项目的所有导入任务。",
        "operationId": "list_import_jobs_api_v1_projects__project_id__import_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}": {
      "get": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Get Import Job",
        "description": "获取导入任务详情和进度。对应 7.2 轮询导入进度。",
        "operationId": "get_import_job_api_v1_projects__project_id__import__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/phase1": {
      "post": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Run Phase1",
        "description": "执行 Phase 1 全量四库分析（异步）。对应 7.3 执行 Phase 1。",
        "operationId": "run_phase1_api_v1_projects__project_id__import__job_id__phase1_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/phase1/result": {
      "get": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Get Phase1 Result",
        "description": "获取 Phase 1 分析结果。",
        "operationId": "get_phase1_result_api_v1_projects__project_id__import__job_id__phase1_result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/phase2": {
      "post": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Run Phase2",
        "description": "执行 Phase 2 近三章动态层分析。对应 7.4 执行 Phase 2。",
        "operationId": "run_phase2_api_v1_projects__project_id__import__job_id__phase2_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/phase2/result": {
      "get": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Get Phase2 Result",
        "description": "获取 Phase 2 分析结果。",
        "operationId": "get_phase2_result_api_v1_projects__project_id__import__job_id__phase2_result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/confirm": {
      "post": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Confirm Import",
        "description": "执行 Phase 3 确认导入（冲突校验 + 事务写入 + 卡牌池生成）。对应 7.5 确认导入。",
        "operationId": "confirm_import_api_v1_projects__project_id__import__job_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "resolve_strategy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "冲突解决策略: keep_existing / merge / replace",
              "default": "keep_existing",
              "title": "Resolve Strategy"
            },
            "description": "冲突解决策略: keep_existing / merge / replace"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/{job_id}/phase3/result": {
      "get": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Get Phase3 Result",
        "description": "获取 Phase 3 导入结果。",
        "operationId": "get_phase3_result_api_v1_projects__project_id__import__job_id__phase3_result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/import/full-import": {
      "post": {
        "tags": [
          "import",
          "Import"
        ],
        "summary": "Full Import",
        "description": "全流程一键导入。\n\n1. 先通过 Phase 0 分章\n2. 自动串联 Phase 1 → Phase 2 → Phase 3\n3. 返回完整导入报告",
        "operationId": "full_import_api_v1_projects__project_id__import_full_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "小说正文纯文本",
              "title": "Text"
            },
            "description": "小说正文纯文本"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "作品名称",
              "default": "",
              "title": "Title"
            },
            "description": "作品名称"
          },
          {
            "name": "resolve_strategy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "冲突解决策略: keep_existing / merge / replace",
              "default": "keep_existing",
              "title": "Resolve Strategy"
            },
            "description": "冲突解决策略: keep_existing / merge / replace"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_import_api_v1_projects__project_id__import_post": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File",
            "description": "上传文件导入"
          }
        },
        "type": "object",
        "title": "Body_submit_import_api_v1_projects__project_id__import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}